requirements.txt
streamlit==1.32.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
altair==5.2.0
scikit-learn==1.3.2
nltk==3.8.1
PyPDF2==3.0.1
python-docx==1.0.1
regex==2023.10.3
docx2txt==0.8
nltk==3.8.1
graphviz==0.20.1
